Add properties to each element
Add a light pink background, and a list of fonts containing `Verdana`
and `DejaVu Sans` with `sans-serif` as a fallback to all odd numbered
elements using class selectors.
Add blue text and a font size of 36px to the second element using an
ID selector.
In addition to the styles for all odd numbered elements, add a font size
of 24px to the third element using multiple classes.
Add a light green background and a font size of 24px, and bold to the
fourth element using an ID selector.